add separate class to check game board
