+++ Gemfile
@@ -1,1 +1,2 @@
+gem 'matrix'
 gem 'rubocop', '~>0.81.0'
